Optimize classifications
There was a big slowdown in testing because the documents
all looked at the entire list of words in the collection,
when almost all of the words didn't exist in the document.
This meant that the term frequencies were 0, which in turn
meant that the overall probability was multiplied by the
some probability raised to the 0 power, which is one. Once
the testing only considered probabilities and term
frequencies for the words in each individual documents, the
speed of training went from more than 30 minutes to less
than a second.

Also, added additional performance tracking.